style: Sonarlint

+++ include/tools/Utils.h
@@ -97,7 +97,7 @@ Coord_i Utils_coord_from_dir_len(Coord_i origin, Direction dir, int len);
  * @param margin margin of error (using double)
  * @return if coord_other is equal coord_mid with a margin
  */
-bool Utils_is_in_middle(Coord_f coord_other, Coord_f coord_mid, float margin);
+bool Utils_is_in_middle(Coord_f coord_other, Coord_f coord_mid, double margin);
 
 /**
  * @brief Return the direction of coord_other from coord_mid with a margin.
@@ -110,7 +110,7 @@ bool Utils_is_in_middle(Coord_f coord_other, Coord_f coord_mid, float margin);
  * @return Direction (Priority: EAST, WEST, SOUTH, NORTH)
  * if nothing is found return NO_DIR (to close to coord_mid)
  */
-Direction Utils_get_dir(Coord_f coord_mid, Coord_f coord_other, float margin);
+Direction Utils_get_dir(Coord_f coord_mid, Coord_f coord_other, double margin);
 
 /**
  * @brief Convert a Coord_i to a Coord_f in the center of the tile.
